+++ src/common/updating/updating.vcxproj
@@ -175,14 +175,14 @@
   <ItemGroup>
     <ClInclude Include="dotnet_installation.h" />
     <ClInclude Include="http_client.h" />
-    <ClInclude Include="toast_notifications_helper.h" />
+    <ClInclude Include="notifications.h" />
     <ClInclude Include="updating.h" />
     <ClInclude Include="pch.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="dotnet_installation.cpp" />
     <ClCompile Include="http_client.cpp" />
-    <ClCompile Include="toast_notifications_helper.cpp" />
+    <ClCompile Include="notifications.cpp" />
     <ClCompile Include="updating.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(CIBuild)'!='true'">Create</PrecompiledHeader>
