+++ src/common/updating/updating.vcxproj.filters
@@ -21,13 +21,13 @@
     <ClInclude Include="updating.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="toast_notifications_helper.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="http_client.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="dotnet_installation.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="notifications.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
@@ -38,13 +38,13 @@
     <ClCompile Include="updating.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="toast_notifications_helper.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="http_client.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="dotnet_installation.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="notifications.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
